refactor(StylePicker): split into StyleStatus and StylePickers
Simplify animation logic by separating the always-visible indicators
from the options panel:

- StyleStatus: fixed indicators showing current selections, click to open
- StylePickers: options panel with simple CSS slide-up animation

Changes:
- Replace complex framer-motion clipPath animations with CSS transform
- Use grid layout for aligned rows with fixed-width label column
- Add gap between pickers and status components
- Square option cells with consistent sizing
