--- src/components/StylePicker/StylePicker.css
@@ -1,93 +1,0 @@
-.StylePicker {
-  /* Fluid sizes: min (mobile), preferred (viewport-based), max (desktop) */
-  --picker-cell: clamp(36px, 8vw, 48px);
-  --picker-gap: clamp(10px, 2vw, 16px);
-  --picker-padding: clamp(10px, 2vw, 16px);
-  --picker-offset: clamp(8px, 2vw, 12px);
-  --picker-font: var(--font-size-fluid-0);
-
-  position: fixed;
-  bottom: calc(var(--picker-offset) + env(safe-area-inset-bottom, 0px));
-  right: var(--picker-offset);
-  z-index: 9999;
-
-  display: flex;
-  flex-direction: column;
-  padding: var(--picker-padding);
-  gap: var(--picker-gap);
-
-  background: var(--bg-dots);
-  box-shadow: var(--shadow-hard-5);
-  border: var(--border-size-2) solid var(--fg-accent);
-}
-
-/* Each row: options + indicator (row-reverse so indicator is on right) */
-.StylePicker__row {
-  display: flex;
-  flex-direction: row-reverse;
-  align-items: center;
-  gap: var(--picker-gap);
-  height: var(--picker-cell);
-
-  background: var(--bg-main);
-}
-
-/* Indicator button (always visible) */
-.StylePicker__indicator {
-  flex-shrink: 0;
-
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  width: var(--picker-cell);
-  height: var(--picker-cell);
-  background: var(--bg-main);
-
-  border: var(--border-size-2) solid var(--fg-accent);
-}
-
-/* Options container (animated) */
-.StylePicker__options {
-  flex-shrink: 0;
-
-  display: grid;
-  grid-auto-flow: column;
-  grid-auto-columns: var(--picker-cell);
-  place-items: center;
-  gap: calc(var(--picker-gap) / 2);
-
-  height: var(--picker-cell);
-}
-
-.StylePicker__option {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  height: 100%;
-  width: 100%;
-
-  border-left: var(--border-size-2) dashed var(--fg-accent);
-  border-right: var(--border-size-2) dashed var(--fg-accent);
-}
-
-/* Row label (visible when expanded) */
-.StylePicker__label {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding-inline: calc(var(--picker-gap) / 2);
-  height: 100%;
-  width: 100%;
-
-  background: var(--fg-accent);
-  color: var(--bg-main);
-  font-size: var(--picker-font);
-  font-weight: var(--font-weight-9);
-
-  text-transform: uppercase;
-  letter-spacing: 0.05em;
-
-  border: var(--border-size-2) dashed var(--fg-accent);
-}
