+++ src/components/StylePicker/StylePickers.css
@@ -1,0 +1,81 @@
+.StylePickers {
+  --picker-cell: clamp(36px, 8vw, 48px);
+  --picker-gap: clamp(10px, 2vw, 16px);
+  --picker-padding: clamp(10px, 2vw, 16px);
+  --picker-offset: clamp(8px, 2vw, 12px);
+  --picker-font: var(--font-size-fluid-0);
+  --picker-label-width: clamp(80px, 20vw, 110px);
+
+  /* Calculate StyleStatus width for positioning */
+  --status-width: calc(
+    var(--picker-cell) + (var(--picker-padding) * 2) + (var(--border-size-2) * 2)
+  );
+  --status-gap: var(--picker-gap);
+
+  position: fixed;
+  /* Position to the left of StyleStatus with gap */
+  bottom: calc(var(--picker-offset) + env(safe-area-inset-bottom, 0px));
+  right: calc(var(--picker-offset) + var(--status-width) + var(--status-gap));
+  z-index: 9999;
+
+  display: flex;
+  flex-direction: column;
+  padding: var(--picker-padding);
+  gap: var(--picker-gap);
+
+  background: var(--bg-dots);
+  box-shadow: var(--shadow-hard-5);
+  border: var(--border-size-2) solid var(--fg-accent);
+
+  /* Slide animation */
+  transform: translateY(calc(100% + var(--picker-offset) * 2));
+  opacity: 0;
+  visibility: hidden;
+  transition:
+    transform 0.2s ease-out,
+    opacity 0.15s ease-out,
+    visibility 0.2s;
+}
+
+.StylePickers--open {
+  transform: translateY(0);
+  opacity: 1;
+  visibility: visible;
+}
+
+.StylePickers__row {
+  display: grid;
+  grid-template-columns: var(--picker-label-width) repeat(3, var(--picker-cell));
+  align-items: center;
+  gap: calc(var(--picker-gap) / 2);
+  height: var(--picker-cell);
+  background: var(--bg-main);
+}
+
+.StylePickers__label {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding-inline: var(--picker-gap);
+  height: 100%;
+
+  background: var(--fg-accent);
+  color: var(--bg-main);
+  font-size: var(--picker-font);
+  font-weight: var(--font-weight-9);
+
+  text-transform: uppercase;
+  letter-spacing: 0.05em;
+
+  border: var(--border-size-2) dashed var(--fg-accent);
+}
+
+.StylePickers__option {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: var(--picker-cell);
+  height: var(--picker-cell);
+  border-left: var(--border-size-2) dashed var(--fg-accent);
+  border-right: var(--border-size-2) dashed var(--fg-accent);
+}
