+++ src/components/StylePicker/StyleStatus.css
@@ -1,0 +1,33 @@
+.StyleStatus {
+  --picker-cell: clamp(36px, 8vw, 48px);
+  --picker-gap: clamp(10px, 2vw, 16px);
+  --picker-padding: clamp(10px, 2vw, 16px);
+  --picker-offset: clamp(8px, 2vw, 12px);
+
+  position: fixed;
+  bottom: calc(var(--picker-offset) + env(safe-area-inset-bottom, 0px));
+  right: var(--picker-offset);
+  z-index: 10000;
+
+  display: flex;
+  flex-direction: column;
+  padding: var(--picker-padding);
+  gap: var(--picker-gap);
+
+  background: var(--bg-dots);
+  box-shadow: var(--shadow-hard-5);
+  border: var(--border-size-2) solid var(--fg-accent);
+
+  cursor: pointer;
+}
+
+
+.StyleStatus__indicator {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: var(--picker-cell);
+  height: var(--picker-cell);
+  background: var(--bg-main);
+  border: var(--border-size-2) solid var(--fg-accent);
+}
