+++ src/components/StylePicker/index.ts
@@ -1,1 +1,2 @@
-export { default } from "./StylePicker";
+export { default as StyleStatus } from "./StyleStatus";
+export { default as StylePickers } from "./StylePickers";
