+++ src/routes/__root.tsx
@@ -8,7 +8,7 @@ import {
   stripSearchParams
 } from "@tanstack/react-router";
 import { TanStackRouterDevtoolsPanel } from "@tanstack/react-router-devtools";
-import StylePicker from "../components/StylePicker";
+import { StylePickers, StyleStatus } from "../components/StylePicker";
 import NotFound from "../components/NotFound";
 import globalCss from "../styles/global.css?url";
 import {
@@ -189,7 +189,8 @@ function RootDocument({ children }: { children: React.ReactNode }) {
         {/* <Header />  faded out until completed*/}
         <main className="BaseLayout">{children}</main>
         <ClientOnly>
-          <StylePicker />
+          <StyleStatus />
+          <StylePickers />
         </ClientOnly>
         <TanStackDevtools
           config={{
